# storage/api-docs/api-docs.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "Almacen API",
        "description": "API para administrar items, categorias, proveedores, ubicaciones y movimientos.",
        "version": "1.0.0"
    },
    "paths": {
...
        "/api/items": {
            "get": {
                "tags": [
                    "Items"
                ],
                "summary": "List items",
                "description": "Display a listing of the resource.",
                "operationId": "2cbf07eb95d033f8775c36424ac8fa06",
                "responses": {
                    "200": {
                        "description": "Items retrieved successfully"
                    }
                }
            },
            "post": {
                "tags": [
                    "Items"
                ],
                "summary": "Create an item",
                "description": "Store a newly created resource in storage.",
                "operationId": "af37857663e6805dd367a55f374c9ee1",
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "required": [
                                    "name",
                                    "price",
                                    "quantity",
                                    "category_id",
                                    "supplier_id",
                                    "location_id"
                                ],
                                "properties": {
                                    "name": {
                                        "type": "string",
                                        "maxLength": 255
                                    },
                                    "description": {
                                        "type": "string",
                                        "nullable": true
                                    },
                                    "price": {
                                        "type": "number",
                                        "format": "float",
                                        "minimum": 0
                                    },
                                    "quantity": {
                                        "type": "integer",
                                        "minimum": 0
                                    },
                                    "category_id": {
                                        "type": "integer"
                                    },
                                    "supplier_id": {
                                        "type": "integer"
                                    },
                                    "location_id": {
                                        "type": "integer"
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "201": {
                        "description": "Item created successfully"
                    },
                    "422": {
                        "description": "Validation error"
                    }
                }
            }
        },
        "/api/items/{item}": {
            "get": {
                "tags": [
                    "Items"
                ],
                "summary": "Show an item",
                "description": "Display the specified resource.",
                "operationId": "95bc89d9f08bcbdb9aa967e2fb4e50a3",
                "parameters": [
                    {
                        "name": "item",
                        "in": "path",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Item retrieved successfully"
                    },
                    "404": {
                        "description": "Item not found"
                    }
                }
            },
            "put": {
                "tags": [
                    "Items"
                ],
                "summary": "Update an item",
                "description": "Update the specified resource in storage.",
                "operationId": "a1d16c822c8ec1a2f4bfd2f968d0c37b",
                "parameters": [
                    {
                        "name": "item",
                        "in": "path",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "properties": {
                                    "name": {
                                        "type": "string",
                                        "maxLength": 255
                                    },
                                    "description": {
                                        "type": "string",
                                        "nullable": true
                                    },
                                    "price": {
                                        "type": "number",
                                        "format": "float",
                                        "minimum": 0
                                    },
                                    "quantity": {
                                        "type": "integer",
                                        "minimum": 0
                                    },
                                    "category_id": {
                                        "type": "integer"
                                    },
                                    "supplier_id": {
                                        "type": "integer"
                                    },
                                    "location_id": {
                                        "type": "integer"
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "Item updated successfully"
                    },
                    "422": {
                        "description": "Validation error"
                    }
                }
            },
            "delete": {
                "tags": [
                    "Items"
                ],
                "summary": "Delete an item",
                "description": "Remove the specified resource from storage.",
                "operationId": "807303f40c62c5a13e0cf9d861d6a0b5",
                "parameters": [
                    {
                        "name": "item",
                        "in": "path",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Item deleted successfully"
                    }
                }
            }
        },
...
    },
...
}